Separated exercise details into components
