+++ src/components/bar-chart/bar-chart.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import  { StatsBarChart } from '../../models/item';
 
@@ -18,7 +18,80 @@ import * as d3Axis from "d3-axis";
 })
 export class BarChartComponent {
 
+	width: number;
+  height: number;
+  margin = {top: 20, right: 20, bottom: 80, left: 0};
+  x: any;
+  y: any;
+  svg: any;
+  g: any;
+
   constructor(public navCtrl: NavController) {
+  	this.width = 1000 - this.margin.left - this.margin.right;
+    this.height = 500 - this.margin.top - this.margin.bottom;
   }
+
+  public makeChart() {
+  	this.initSvg()
+    this.initAxis();
+    this.drawAxis();
+    this.drawBars();
+  }
+
+  initSvg() {
+
+    this.svg = d3.select("#barChart")
+        .append("svg")
+        .attr("width", '100%')
+        .attr("height", '100%')
+        .attr('viewBox','0 0 900 500');
+    this.g = this.svg.append("g")
+        .attr("transform", "translate(" + this.margin.left + "," + this.margin.top + ")");
+	}
+
+	initAxis() {
+    this.x = d3Scale.scaleBand().rangeRound([0, this.width]).padding(0.1);
+    this.y = d3Scale.scaleLinear().rangeRound([this.height, 0]);
+    this.x.domain(StatsBarChart.map((d) => d.reps));
+    this.y.domain([0, d3Array.max(StatsBarChart, (d) => d.oneRM)]);
+  }
+
+  drawAxis() {
+    this.g.append("g")
+        .attr("class", "axis axis--x")
+        .attr("transform", "translate(0," + this.height + ")")
+        .call(d3Axis.axisBottom(this.x))
+        .append("text")
+        .attr("class", "axis-title")
+        .attr("y", 70)
+        .attr("x", this.width / 2)
+        .attr("text-anchor", "end")
+        .text("Reps");
+    this.g.append("g")
+        .attr("class", "axis axis--y")
+        .call(d3Axis.axisLeft(this.y))
+        .append("text")
+        .attr("class", "axis-title")
+        .attr("transform", "rotate(-90)")
+        .attr("y", -90)
+        .attr("x", (this.height / -2) + 20)
+        .attr("dy", "0.71em")
+        .attr("text-anchor", "end")
+        .text("1RM");
+  }
+
+  drawBars() {
+    this.g.selectAll(".bar")
+        .data(StatsBarChart)
+        .enter().append("rect")
+        .attr("class", "bar")
+        .attr("x", (d) => this.x(d.reps) )
+        .attr("y", (d) => this.y(d.oneRM) )
+        .attr("width", this.x.bandwidth())
+        .attr("height", (d) => this.height - this.y(d.oneRM) );
+  }
+
+
+
 
 }
