+++ src/components/history/history.html
@@ -2,18 +2,18 @@
 <div>
   <ion-item>
     <div class="scrollingtable">
-    	<table id="headers">
-    		<tr>
-    			<th>Date</th>
-    			<th>Weight</th>
-    			<th>Reps</th>
-    			<th>1RM</th>
-    			<th>Delete</th>
-    		</tr>
-    	</table>
+      <table id="headers">
+        <tr>
+          <th>Date</th>
+          <th>Weight</th>
+          <th>Reps</th>
+          <th>1RM</th>
+          <th>Delete</th>
+        </tr>
+      </table>
       <ion-scroll scrollY="true" class="scroll">
         <table>
-          <tr *ngFor="let x of history">
+          <tr *ngFor="let x of history._history">
             <td>{{x.date | date: 'MM/dd'}}</td>
             <td>{{x.weight}}</td>
             <td>{{x.reps}}</td>
