+++ src/components/history/history.scss
@@ -17,6 +17,10 @@ history {
 		width: 200px;
 	}
 
+	.scrollingtable th {
+		width: 200px;
+	}
+
 	.remove {
 		background-color: red;
 	}
