+++ src/components/line-chart/line-chart.ts
@@ -1,4 +1,12 @@
 import { Component } from '@angular/core';
+import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import  { StatsLineChart } from '../../models/item';
+
+import * as d3 from 'd3-selection';
+import * as d3Scale from "d3-scale";
+import * as d3Shape from "d3-shape";
+import * as d3Array from "d3-array";
+import * as d3Axis from "d3-axis";
 
 /**
  * Generated class for the LineChartComponent component.
@@ -12,7 +20,79 @@ import { Component } from '@angular/core';
 })
 export class LineChartComponent {
 
+	width2: number;
+  height2: number;
+  margin2 = {top: 20, right: 20, bottom: 80, left: 0};
+  x2: any;
+  y2: any;
+  svg2: any;
+  g2: any;
+
+  line: d3Shape.Line<[number, number]>;
+
   constructor() {
+  	
+  	this.width2 = 1000 - this.margin2.left - this.margin2.right;
+    this.height2 = 500 - this.margin2.top - this.margin2.bottom;
+  }
+
+  public makeChart2() {
+  	this.initSvg()
+    this.initAxis();
+    this.drawAxis();
+    this.drawLine();
+  }
+
+  initSvg() {
+    this.svg2 = d3.select("#lineChart")
+        .append("svg")
+        .attr("width", '100%')
+        .attr("height", '100%')
+        .attr('viewBox','0 0 900 500');
+    this.g2 = this.svg2.append("g")
+        .attr("transform", "translate(" + this.margin2.left + "," + this.margin2.top + ")");
+  }
+
+  initAxis() {
+    this.x2 = d3Scale.scaleBand().rangeRound([0, this.width2]).padding(0.1);
+    this.y2 = d3Scale.scaleLinear().rangeRound([this.height2, 0]);
+    this.x2.domain(StatsLineChart.map((d) => d.date));
+    this.y2.domain([0, d3Array.max(StatsLineChart, (d) => d.oneRM)]);
+  }
+
+  drawAxis() {
+    this.g2.append("g")
+        .attr("class", "axis axis--x")
+        .attr("transform", "translate(0," + this.height2 + ")")
+        .call(d3Axis.axisBottom(this.x2))
+        .append("text")
+        .attr("class", "axis-title")
+        .attr("y", 70)
+        .attr("x", this.width2 / 2)
+        .attr("text-anchor", "end")
+        .text("Date");
+    this.g2.append("g")
+        .attr("class", "axis axis--y")
+        .call(d3Axis.axisLeft(this.y2))
+        .append("text")
+        .attr("class", "axis-title")
+        .attr("transform", "rotate(-90)")
+        .attr("y", -90)
+        .attr("x", (this.height2 / -2) + 20)
+        .attr("dy", "0.71em")
+        .attr("text-anchor", "end")
+        .text("1RM");
+  }
+
+  drawLine() {
+    this.line = d3Shape.line()
+        .x( (d: any) => this.x2(d.date) )
+        .y( (d: any) => this.y2(d.oneRM) );
+
+    this.g2.append("path")
+        .datum(StatsLineChart)
+        .attr("class", "line")
+        .attr("d", this.line);
   }
 
 }
