+++ src/components/new-set/new-set.html
@@ -1,20 +1,25 @@
 <!-- Generated template for the NewSetComponent component -->
 <div>
   <ion-item>
-		<div class="levels">
-			<h2>LEVEL <span class="sick"> {{xlevel}}</span> : <span class="sick">{{xcurrent}}</span> / <span class="sick">{{xtotal}}</span> GAINS</h2>
-		</div>
-		<div id="myProgress">
+    	<div class="levels">
+    		<h2>LEVEL <span class="sick"> {{xlevel}}</span> : <span class="sick">{{xcurrent}}</span> / <span class="sick">{{xtotal}}</span> GAINS</h2>
+    	</div>
+    	<div id="myProgress">
 			<div id="myBar" [ngStyle]="{'width.%' : progress}"></div>
 		</div>
 		<div class="breaker"></div>
-		<div class="setbet">
-			<h2>Weight: <input type="text" value="100"></h2>
-			<br>
-			<h2>Reps: <input type="text" value="10"></h2>
-			<br>
-			<h2><input type="submit" id="add" value="+"></h2>
-			<br>
-		</div>
-	</ion-item>
+    	<div class="setbet">
+    		<h2>Weight: <input [(ngModel)]="weight" type="text" value="100"></h2>
+    		<br>
+    		<h2>Reps: <input [(ngModel)]="reps" type="text" value="10"></h2>
+    		<br>
+    		<h2><input (click)="addSet()" type="submit" id="add" value="+"></h2>
+    		<br>
+    	</div>
+    	<br>
+    	<br>
+    	<div *ngIf="bool" class="animation">
+    		<h1>NEW RECORD!</h1>
+    	</div>
+    </ion-item>
 </div>
