+++ src/components/new-set/new-set.scss
@@ -41,4 +41,9 @@ new-set {
 	  margin-bottom : 10px;
 	  margin-top: 5px;
 	}
+
+  .animation {
+    text-align: center;
+    animation: bounceInUp 1s;
+  }
 }
