+++ src/components/records/records.html
@@ -1,24 +1,21 @@
 <!-- Generated template for the RecordsComponent component -->
 <div>
   <ion-item>
-  	<div class="rectable">
+    <div class="rectable">
       <table>
-      	<tr>
-    			<th>Reps</th>
-    			<th>Weight</th>
-    			<th>1RM</th>
-    			<th>Records</th>
-    		</tr>
-    		<tr *ngFor="let x of records" [ngClass]="{red: x.records == 0, green: x.records > 0}">
-    			<td class="bold">{{x.reps}}</td>
-    			<td>{{x.weight}}</td>
-    			<td>{{x.oneRM}}</td>
-    			<td>{{x.records}}</td>
-    		</tr>
-    	  </table>
-    	</div>
-    	<!--Here is where I want the bar graph but I CANNOT see it
-    	<svg id="barChart"></svg>
-    	-->
+        <tr>
+          <th>Reps</th>
+          <th>Weight</th>
+          <th>1RM</th>
+          <th>Records</th>
+        </tr>
+        <tr *ngFor="let x of records._records" [ngClass]="{red: x.records == 0, green: x.records > 0 && x.records < 5, silver: x.records > 4 && x.records < 10, gold: x.records > 9 }">
+          <td class="bold">{{x.reps}}</td>
+          <td>{{x.weight}}</td>
+          <td>{{x.oneRM.toFixed(1)}}</td>
+          <td>{{x.records}}</td>
+        </tr>
+        </table>
+      </div>
   </ion-item>
 </div>
