+++ src/pages/exercise-detail/exercise-detail.html
@@ -24,78 +24,15 @@
 
 	<div [ngSwitch]="segment">
 	  <ion-list *ngSwitchCase="'set'">
-	    <ion-item>
-	    	<div class="levels">
-	    		<h2>LEVEL <span class="sick"> {{xlevel}}</span> : <span class="sick">{{xcurrent}}</span> / <span class="sick">{{xtotal}}</span> GAINS</h2>
-	    	</div>
-	    	<div id="myProgress">
-				<div id="myBar" [ngStyle]="{'width.%' : progress}"></div>
-			</div>
-			<div class="breaker"></div>
-	    	<div class="setbet">
-	    		<h2>Weight: <input [(ngModel)]="weight" type="text" value="100"></h2>
-	    		<br>
-	    		<h2>Reps: <input [(ngModel)]="reps" type="text" value="10"></h2>
-	    		<br>
-	    		<h2><input (click)="addSet()" type="submit" id="add" value="+"></h2>
-	    		<br>
-	    	</div>
-	    	<br>
-	    	<br>
-	    	<div *ngIf="bool" class="animation">
-	    		<h1>NEW RECORD!</h1>
-	    	</div>
-	    </ion-item>
+	    <new-set></new-set>
 	  </ion-list>
 
 	  <ion-list *ngSwitchCase="'history'">
-<ion-item>
-    <div class="scrollingtable">
-    	<table id="headers">
-    		<tr>
-    			<th>Date</th>
-    			<th>Weight</th>
-    			<th>Reps</th>
-    			<th>1RM</th>
-    			<th>Delete</th>
-    		</tr>
-    	</table>
-      <ion-scroll scrollY="true" class="scroll">
-        <table>
-          <tr *ngFor="let x of history">
-            <td>{{x.date | date: 'MM/dd'}}</td>
-            <td>{{x.weight}}</td>
-            <td>{{x.reps}}</td>
-            <td>{{x.oneRM.toFixed(1)}}</td>
-            <td>
-              <input type="button" value="X" class="remove">
-            </td>
-          </tr>
-        </table>
-      </ion-scroll>
-    </div>
-  </ion-item>
+	  	<history></history>
 	  </ion-list>
 
 	  <ion-list *ngSwitchCase="'records'">
-	    <ion-item>
-	    	<div class="rectable">
-		      <table>
-		      	<tr>
-	      			<th>Reps</th>
-	      			<th>Weight</th>
-	      			<th>1RM</th>
-	      			<th>Records</th>
-	      		</tr>
-	      		<tr *ngFor="let x of records._records" [ngClass]="{red: x.records == 0, green: x.records > 0 && x.records < 5, silver: x.records > 4 && x.records < 10, gold: x.records > 9 }">
-	      			<td class="bold">{{x.reps}}</td>
-	      			<td>{{x.weight}}</td>
-	      			<td>{{x.oneRM.toFixed(1)}}</td>
-	      			<td>{{x.records}}</td>
-	      		</tr>
-	      	  </table>
-	      	</div>
-	    </ion-item>
+	    <records></records>
 	  </ion-list>
 
 	  <bar-chart [hidden]="selectedValue!=1" ></bar-chart>
