+++ src/pages/exercise-detail/exercise-detail.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import  { StatsBarChart } from '../../models/item';
 import  { StatsLineChart } from '../../models/item';
@@ -7,12 +7,12 @@ import { Items } from '../../providers/providers';
 import { Records } from '../../providers/providers';
 import { User } from '../../providers/providers';
 import { Levels } from '../../providers/providers';
-//import { History } from '../../providers/providers';
+import { HistoryProvider } from '../../providers/providers';
 
 import firebase from 'firebase';
 
 import { BarChartComponent } from '../../components/bar-chart/bar-chart';
-
+import { LineChartComponent } from '../../components/line-chart/line-chart';
 
 import * as d3 from 'd3-selection';
 import * as d3Scale from "d3-scale";
@@ -41,7 +41,7 @@ export class ItemDetailPage {
   loop = 0;
   bool = false;
 
-  history = [{date: 1, reps: 1, weight: 1, oneRM: 1}];
+  //history = [{date: 1, reps: 1, weight: 1, oneRM: 1}];
   myRecords = [{reps: 1, weight: 1, oneRM: 1, records: 1}];
 
   item: any;
@@ -68,12 +68,16 @@ export class ItemDetailPage {
 
   line: d3Shape.Line<[number, number]>;
 
+  @ViewChild(BarChartComponent) barChart: BarChartComponent
+  @ViewChild(LineChartComponent) lineChart: LineChartComponent
+
   constructor(public navCtrl: NavController,
     navParams: NavParams,
     items: Items,
     public user: User,
     private records: Records,
-    public levels: Levels) {
+    public levels: Levels,
+    private history: HistoryProvider) {
 
     this.exercise = navParams.get('item') || items.defaultItem;
     this.item = navParams.get('item') || items.defaultItem;
@@ -83,63 +87,16 @@ export class ItemDetailPage {
     this.height2 = 500 - this.margin2.top - this.margin2.bottom;
   }
 
-  ionViewDidLoad() {
+  ionViewDidEnter() {
     this.username = this.user._user
     this.myRecords = [];
-    this.history = [];
-
-    var queryHistory = firebase.database().ref('/' + this.username + '/exercises/' + this.exercise.name + '/history');
-    queryHistory.once("value").then( snapshot => {
-      snapshot.forEach( childSnapshot => {
-        var childData1 = childSnapshot.val();
-        this.history.push(childData1);     
-      });
-    });
+    this.history._history = [];
     
-    var queryGains = firebase.database().ref('/' + this.username + '/gains');
-    queryGains.once("value").then( snapshot => {
-      this.loop = 0;
-      this.gains = 0;
-      snapshot.forEach( childSnapshot => {
-        this.loop++
-        var childData2 = childSnapshot.val();
-        var gains = childData2.gains;
-        this.gains = this.gains + gains
-        if ( snapshot.numChildren() == this.loop )
-          this.setLevel()
-      })
-    })
-
-    var count = 0; 
-    var queryRecords = firebase.database().ref('/' + this.username + '/exercises/' + this.exercise.name + '/records');
-    queryRecords.once("value").then( snapshot => {
-      snapshot.forEach( childSnapshot => {
-        var childData1 = childSnapshot.val();
-        var r = {reps: childData1.reps, weight: childData1.weight, oneRM: childData1.oneRM, records: childData1.records};
-        this.records._records[count] = r;
-        count++     
-      });
-    }); 
-
-    this.initSvg()
-    this.initAxis();
-    this.drawAxis();
-    this.drawBars();
-    this.drawLine();
+    this.barChart.makeChart();
+    this.lineChart.makeChart2();
 
   }
-
-  setLevel () {
-    this.levels._levels.forEach( ( value, index) => {
-      if (this.gains > value.totalPoints) {
-        this.xcurrent = this.gains - value.totalPoints;
-        this.xlevel = value.level;
-        this.xtotal = value.levelPoints;
-        this.progress = this.xcurrent / this.xtotal * 100
-      }
-    });
-  }
-
+  
   showBar() {
     this.selectedValue = 1;
   }
@@ -151,140 +108,5 @@ export class ItemDetailPage {
   hideCharts() {
     this.selectedValue = 0;
   }
-
-  initSvg() {
-
-    this.svg = d3.select("#barChart")
-        .append("svg")
-        .attr("width", '100%')
-        .attr("height", '100%')
-        .attr('viewBox','0 0 900 500');
-    this.g = this.svg.append("g")
-        .attr("transform", "translate(" + this.margin.left + "," + this.margin.top + ")");
-
-    this.svg2 = d3.select("#lineChart")
-        .append("svg")
-        .attr("width", '100%')
-        .attr("height", '100%')
-        .attr('viewBox','0 0 900 500');
-    this.g2 = this.svg2.append("g")
-        .attr("transform", "translate(" + this.margin2.left + "," + this.margin2.top + ")");
-  }
-
-  initAxis() {
-    this.x = d3Scale.scaleBand().rangeRound([0, this.width]).padding(0.1);
-    this.y = d3Scale.scaleLinear().rangeRound([this.height, 0]);
-    this.x.domain(this.records._records.map((d) => d.reps));
-    this.y.domain([0, d3Array.max(this.records._records, (d) => d.oneRM)]);
-
-    this.x2 = d3Scale.scaleBand().rangeRound([0, this.width2]).padding(0.1);
-    this.y2 = d3Scale.scaleLinear().rangeRound([this.height2, 0]);
-    this.x2.domain(this.history.map((d) => d.date));
-    this.y2.domain([0, d3Array.max(this.history, (d) => d.oneRM)]);
-  }
-
-  drawAxis() {
-    this.g.append("g")
-        .attr("class", "axis axis--x")
-        .attr("transform", "translate(0," + this.height + ")")
-        .call(d3Axis.axisBottom(this.x))
-        .append("text")
-        .attr("class", "axis-title")
-        .attr("y", 70)
-        .attr("x", this.width / 2)
-        .attr("text-anchor", "end")
-        .text("Reps");
-    this.g.append("g")
-        .attr("class", "axis axis--y")
-        .call(d3Axis.axisLeft(this.y))
-        .append("text")
-        .attr("class", "axis-title")
-        .attr("transform", "rotate(-90)")
-        .attr("y", -90)
-        .attr("x", (this.height / -2) + 20)
-        .attr("dy", "0.71em")
-        .attr("text-anchor", "end")
-        .text("1RM");
-
-    this.g2.append("g")
-        .attr("class", "axis axis--x")
-        .attr("transform", "translate(0," + this.height2 + ")")
-        .call(d3Axis.axisBottom(this.x2))
-        .append("text")
-        .attr("class", "axis-title")
-        .attr("y", 70)
-        .attr("x", this.width2 / 2)
-        .attr("text-anchor", "end")
-        .text("Date");
-    this.g2.append("g")
-        .attr("class", "axis axis--y")
-        .call(d3Axis.axisLeft(this.y2))
-        .append("text")
-        .attr("class", "axis-title")
-        .attr("transform", "rotate(-90)")
-        .attr("y", -90)
-        .attr("x", (this.height2 / -2) + 20)
-        .attr("dy", "0.71em")
-        .attr("text-anchor", "end")
-        .text("1RM");
-  }
-
-  drawBars() {
-    this.g.selectAll(".bar")
-        .data(this.records._records)
-        .enter().append("rect")
-        .attr("class", "bar")
-        .attr("x", (d) => this.x(d.reps) )
-        .attr("y", (d) => this.y(d.oneRM) )
-        .attr("width", this.x.bandwidth())
-        .attr("height", (d) => this.height - this.y(d.oneRM) );
-  }
-
-  drawLine() {
-    this.line = d3Shape.line()
-        .x( (d: any) => this.x2(d.date) )
-        .y( (d: any) => this.y2(d.oneRM) );
-
-    this.g2.append("path")
-        .datum(this.history)
-        .attr("class", "line")
-        .attr("d", this.line);
-  }
-
-  addSet() {
-    var date = new Date().toISOString();
-    var oneRM = this.weight / (1.0278- (this.reps * .0278));
-    var gains = 5;
-    this.bool = false;
-
-    this.records._records.forEach( (value, index) => {
-      if (this.reps == value.reps) {
-        if (this.weight > value.weight) {
-          this.records._records[index].weight = this.weight;
-          this.records._records[index].oneRM = oneRM;
-          this.records._records[index].records++;
-          gains = 10;
-          this.bool = true;
-        }
-      }
-    });
-
-    setTimeout(() => {
-      this.bool = false;
-    }, 2000);
-
-    var set = { date: date, weight: this.weight, reps: this.reps, oneRM: oneRM, gains: gains};
-    var g = { date: date, gains: gains};
-    
-    var history = firebase.database().ref('/' + this.username + '/exercises/' + this.exercise.name + '/history');
-    history.push(set);
-
-    var points = firebase.database().ref('/' + this.username + '/gains');
-    points.push(g);
-
-    var records = firebase.database().ref('/' + this.username + '/exercises/' + this.exercise.name + '/records');
-    records.set(this.records._records);
-
-    this.ionViewDidLoad();
-  }
+  
 }
